refactor: reorganize image handling services
- Moved image-related functions from services/fetch to lib/images/actions for better structure and maintainability.
- Updated imports across various components to reflect the new location of image handling functions.
- Removed obsolete API routes and types related to image uploads to streamline the codebase.
