+++ app/admin/add-barometer/file-upload.tsx
@@ -10,7 +10,7 @@ import { Card } from '@/components/ui/card'
 import { FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form'
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip'
 import { imageStorage } from '@/constants/globals'
-import { createImageUrls, deleteImage, uploadFileToCloud } from '@/services/fetch'
+import { createImageUrls, deleteImage, uploadFileToCloud } from '@/lib/images/actions'
 
 interface FileUploadProps {
   name: string
