--- app/api/v2/parameters.ts
@@ -1,2 +1,0 @@
-// pagination page size
-export const DEFAULT_PAGE_SIZE = 12
