--- app/api/v2/upload/images/types.ts
@@ -1,4 +1,0 @@
-export type FileProps = { fileName: string; contentType: string }
-export type FileDto = { files: FileProps[] }
-export type UrlProps = { signed: string; public: string }
-export type UrlDto = { urls: UrlProps[] }
