+++ app/brands/brand-image-edit.tsx
@@ -13,7 +13,7 @@ import NextImage from 'next/image'
 import { type TransitionStartFunction, useCallback } from 'react'
 import type { UseFormReturn } from 'react-hook-form'
 import { Button } from '@/components/ui/button'
-import { createImageUrls, deleteImage, uploadFileToCloud } from '@/services/fetch'
+import { createImageUrls, deleteImage, uploadFileToCloud } from '@/lib/images/actions'
 import type { BrandEditForm } from './brand-edit-schema'
 
 interface Props {
