+++ services/index.ts
@@ -1,1 +1,3 @@
-export * from './fetch'
+export * from './auth'
+export * from './cookie-consent'
+export * from './minio'
