--- types/api.ts
@@ -1,1 +1,0 @@
-export type { FileProps, UrlDto } from '../app/api/v2/upload/images/types'
