+++ types/index.ts
@@ -1,3 +1,2 @@
-export * from './api'
 export * from './models'
 export * from './next'
